Port to macOS/POSIX: first native build that compiles its own game module
Fixes, in order of how much they matter:

* inc/Defines.h: int32/uint32/rID/hObj were typedef'd to long. Windows is
  LLP64 so long is 4 bytes there, but macOS and Linux are LP64 where it is 8.
  Every 32-bit type in the codebase was silently 64-bit. The codebase catches
  this itself -- inc/RComp.h:221 asserts sizeof(int32) <= sizeof(VCode), and
  that assert fires on every opcode emitted. Narrowed them to int/unsigned int.

* Consequences of the above: YYSTYPE was long (inc/yygram.h); several parser
  forward declarations in lang/Grammar.acc and src/yygram.cpp said long* while
  their definitions said YYSTYPE*, so they only ever matched by coincidence.

* Windows-only constructs: <malloc.h> (see compat/), <io.h> in src/Tokens.cpp,
  and BREAKOUT expanding to the x86 asm 'int $3', which does not assemble on
  arm64. Replaced with __builtin_debugtrap().

* 104 hardcoded backslash path separators across 6 built source files.

* Three sites passing a non-trivial String through printf-style varargs, which
  aborts at runtime under the arm64 calling convention.

Adds build_macos.sh, which builds libtcod from the vendored copy, links against
Homebrew SDL2 and system zlib, and compiles mod/Incursion.Mod from lib/main.irc.

Not yet done: the pdcurses backend (src/Wcurses.cpp) is still Windows-only, so
there is no headless text mode. No x86_64 slice, no Linux verification, no CI.
26 printf format mismatches remain, plus 32 format bugs that are also live on
Windows (wrong arg counts, invalid specifiers).

+++ lib/dispatch.h
@@ -3548,7 +3548,7 @@ void VMachine::SetMemberVar(int16 varid, hObj h, int32 val)
           pe->ESpellNum = (int16)val;
          break;
         case 48:
-          pe->EParam = val;
+          pe->EParam = (int16)val;
          break;
         case 49:
           pe->EDir = (int16)val;
